refactor: move processWeather to its own file
